Isolated the loading of assemblies into a separate AssemblyLoadContext when generating mermaid diagrams from code, to prevent version clashes with our own dependencies
